Pull in crc32 from old varnish sources to remove dependency on libvgz/libz
crc32.c is largely equivalent to the freebsd implementation

Revert "use includes from VARNISHSRC if we need to" -
This reverts commit 496725d6123d36e333134d5a1733c363922cf6f8.

The decision to pull in crc32 is based on an irc dicussion with phk
that vmods should not use libvgz. The libvgz api is not to be exposed
to minimize name clash issues with libz. As we only need crc32, there
is no good reason to use libvgz in the first place.